requirements.txt
# Backend
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=2.0.0

# LLM and orchestration
openai>=1.0.0
langgraph>=0.0.40

# Database
supabase>=2.0.0

# Stats and data
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Frontend
streamlit>=1.28.0
plotly>=5.18.0
